+++ src/app.py
@@ -1,19 +1,18 @@
 import os
-import sys
 import logging
-import requests
-import json
 import requests
 # Don't confuse urllib (Python native library) with urllib3 (3rd-party library, requests also uses urllib3)
 from requests.packages.urllib3.exceptions import InsecureRequestWarning
 import argparse
-from pathlib import Path
 from shutil import rmtree # Used by file removal
 from flask import Flask, jsonify, abort, request, session, redirect, Response
 from globus_sdk import AccessTokenAuthorizer, AuthClient, ConfidentialAppAuthClient
 
 # HuBMAP commons
 from hubmap_commons.hm_auth import AuthHelper
+
+from routes.auth import auth_blueprint
+from routes.status import status_blueprint
 
 # Set logging format and level (default is warning)
 # All the API logging is forwarded to the uWSGI server and gets written into the log file `uwsgi-ingest-api.log`
@@ -25,6 +24,9 @@ logger = logging.getLogger(__name__)
 # Specify the absolute path of the instance folder and use the config file relative to the instance path
 app = Flask(__name__, instance_path=os.path.join(os.path.abspath(os.path.dirname(__file__)), 'instance'), instance_relative_config=True)
 app.config.from_pyfile('app.cfg')
+
+app.register_blueprint(auth_blueprint)
+app.register_blueprint(status_blueprint)
 
 # Suppress InsecureRequestWarning warning when requesting status on https with ssl cert verify disabled
 requests.packages.urllib3.disable_warnings(category = InsecureRequestWarning)
@@ -39,15 +41,18 @@ requests.packages.urllib3.disable_warnings(category = InsecureRequestWarning)
 def http_bad_request(e):
     return jsonify(error=str(e)), 400
 
+
 # Error handler for 401 Unauthorized with custom error message
 @app.errorhandler(401)
 def http_unauthorized(e):
     return jsonify(error=str(e)), 401
 
+
 # Error handler for 404 Not Found with custom error message
 @app.errorhandler(404)
 def http_not_found(e):
     return jsonify(error=str(e)), 404
+
 
 # Error handler for 500 Internal Server Error with custom error message
 @app.errorhandler(500)
@@ -74,176 +79,16 @@ except Exception:
     logger.exception(msg)
 
 
-####################################################################################################
-## Default and Status Routes
-####################################################################################################
-
-@app.route('/', methods = ['GET'])
+# The only endpoint that should be in this file, all others should be route Blueprints...
+@app.route('/', methods=['GET'])
 def index():
     return "Hello! This is SenNet Ingest API service :)"
 
 
-"""
-Show status of the current VERSION and BUILD
-Returns
--------
-json
-    A json containing the status details
-"""
-@app.route('/status', methods = ['GET'])
-def get_status():
-    status_data = {
-        # Use strip() to remove leading and trailing spaces, newlines, and tabs
-        'version': (Path(__file__).absolute().parent.parent / 'VERSION').read_text().strip(),
-        'build': (Path(__file__).absolute().parent.parent / 'BUILD').read_text().strip()
-    }
-
-    return jsonify(status_data)
-
-
-####################################################################################################
-## Endpoints for UI Login and Logout
-####################################################################################################
-
-# Redirect users from react app login page to Globus auth login widget then redirect back
-@app.route('/login')
-def login():
-    #redirect_uri = url_for('login', _external=True)
-    redirect_uri = app.config['FLASK_APP_BASE_URI'] + 'login'
-
-    confidential_app_auth_client = ConfidentialAppAuthClient(app.config['APP_CLIENT_ID'], app.config['APP_CLIENT_SECRET'])
-    confidential_app_auth_client.oauth2_start_flow(redirect_uri, refresh_tokens=True)
-
-    # If there's no "code" query string parameter, we're in this route
-    # starting a Globus Auth login flow.
-    # Redirect out to Globus Auth
-    if 'code' not in request.args:
-        auth_uri = confidential_app_auth_client.oauth2_get_authorize_url(additional_params={"scope": "openid profile email urn:globus:auth:scope:transfer.api.globus.org:all urn:globus:auth:scope:auth.globus.org:view_identities urn:globus:auth:scope:groups.api.globus.org:all"})
-        return redirect(auth_uri)
-    # If we do have a "code" param, we're coming back from Globus Auth
-    # and can start the process of exchanging an auth code for a token.
-    else:
-        auth_code = request.args.get('code')
-
-        token_response = confidential_app_auth_client.oauth2_exchange_code_for_tokens(auth_code)
-
-        # Get all Bearer tokens
-        auth_token = token_response.by_resource_server['auth.globus.org']['access_token']
-        #nexus_token = token_response.by_resource_server['nexus.api.globus.org']['access_token']
-        transfer_token = token_response.by_resource_server['transfer.api.globus.org']['access_token']
-        groups_token = token_response.by_resource_server['groups.api.globus.org']['access_token']
-        # Also get the user info (sub, email, name, preferred_username) using the AuthClient with the auth token
-        user_info = get_user_info(auth_token)
-
-        info = {
-            'name': user_info['name'],
-            'email': user_info['email'],
-            'globus_id': user_info['sub'],
-            'auth_token': auth_token,
-            'transfer_token': transfer_token,
-            'groups_token': groups_token
-        }
-
-        # Turns json dict into a str
-        json_str = json.dumps(info)
-        #print(json_str)
-
-        # Store the resulting tokens in server session
-        session.update(
-            tokens=token_response.by_resource_server
-        )
-
-        # Finally redirect back to the client
-        return redirect(app.config['GLOBUS_CLIENT_APP_URI'] + '?info=' + str(json_str))
-
-
-@app.route('/logout')
-def logout():
-    """
-    - Revoke the tokens with Globus Auth.
-    - Destroy the session state.
-    - Redirect the user to the Globus Auth logout page.
-    """
-    confidential_app_auth_client = ConfidentialAppAuthClient(app.config['APP_CLIENT_ID'], app.config['APP_CLIENT_SECRET'])
-
-    # Revoke the tokens with Globus Auth
-    if 'tokens' in session:
-        for token in (token_info['access_token']
-            for token_info in session['tokens'].values()):
-                confidential_app_auth_client.oauth2_revoke_token(token)
-
-    # Destroy the session state
-    session.clear()
-
-    # build the logout URI with query params
-    # there is no tool to help build this (yet!)
-    globus_logout_url = (
-        'https://auth.globus.org/v2/web/logout' +
-        '?client={}'.format(app.config['APP_CLIENT_ID']) +
-        '&redirect_uri={}'.format(app.config['GLOBUS_CLIENT_APP_URI']) +
-        '&redirect_name={}'.format(app.config['GLOBUS_CLIENT_APP_NAME']))
-
-    # Redirect the user to the Globus Auth logout page
-    return redirect(globus_logout_url)
-
-
-####################################################################################################
-## Ingest API Endpoints
-####################################################################################################
-
+# Should be moved to a routes Blueprint when implemented
 @app.route('/datasets', methods=['POST'])
 def create_dataset():
     return "Placeholder"
-
-
-####################################################################################################
-## Internal Functions
-####################################################################################################
-
-"""
-Throws error for 400 Bad Reqeust with message
-Parameters
-----------
-err_msg : str
-    The custom error message to return to end users
-"""
-def bad_request_error(err_msg):
-    abort(400, description = err_msg)
-
-"""
-Throws error for 401 Unauthorized with message
-Parameters
-----------
-err_msg : str
-    The custom error message to return to end users
-"""
-def unauthorized_error(err_msg):
-    abort(401, description = err_msg)
-
-"""
-Throws error for 404 Not Found with message
-Parameters
-----------
-err_msg : str
-    The custom error message to return to end users
-"""
-def not_found_error(err_msg):
-    abort(404, description = err_msg)
-
-"""
-Throws error for 500 Internal Server Error with message
-Parameters
-----------
-err_msg : str
-    The custom error message to return to end users
-"""
-def internal_server_error(err_msg):
-    abort(500, description = err_msg)
-    
-
-def get_user_info(token):
-    auth_client = AuthClient(authorizer=AccessTokenAuthorizer(token))
-    return auth_client.oauth2_userinfo()
 
 
 ####################################################################################################
