+++ src/requirements.txt
@@ -1,9 +1,9 @@
 Flask==2.1.3
-globus_sdk==2.0.1
+globus-sdk==3.0.2
 hubmap-sdk==1.0.2
 
 # The commons package requires requests>=2.22.0
-requests==2.25.1
+requests==2.26.0
 
 # Use the published package from PyPI as default
 # Use the branch name of commons from github for testing new changes made in commons from different branch
